Grass test work/fix. Grass noise texture frequency. Quads render textures ~size display. All classes that uses Logger now will log in one file. Realtime Atmosphere Baker tweaks/fixes/additions. Quads flicker fix, when rendering from Update(). Planetoid.LODDistanceMult. now float.

+++ Assets/Project/SpaceEngine/Code/Planetoid.cs
@@ -68,7 +68,8 @@ public sealed class Planetoid : MonoBehaviour
 
     public int DispatchSkipFramesCount = 8;
 
-    public int LODDistanceMultiplier = 1;
+    public float LODDistanceMultiplier = 1;
+
     public int LODMaxLevel = 8;
     public int[] LODDistances = new int[11] { 2048, 1024, 512, 256, 128, 64, 32, 16, 8, 4, 2 };
